+++ RUFAS/data_types/max_daily_feed.py
