+++ RUFAS/data_types/total_inventory.py
