+++ RUFAS/routines/animal/animal_manager.py
@@ -31,7 +31,7 @@ from RUFAS.time import Time
 from RUFAS.units import MeasurementUnits
 from RUFAS.weather import Weather
 
-from ...data_structures.pen_manure_data import PenManureData
+from ...data_types.pen_manure_data import PenManureData
 from ...enums import AnimalCombination
 
 om = OutputManager()
