+++ RUFAS/routines/animal/manure/dry_cow_manure_excretion.py
@@ -5,7 +5,7 @@ from RUFAS.general_constants import GeneralConstants
 from RUFAS.routines.animal.animal_module_constants import AnimalModuleConstants
 from RUFAS.routines.animal.manure.general_manure import calculate_phosphorus_excretion_values
 
-from ....data_structures.animal_manure_excretions import AnimalManureExcretions
+from ....data_types.animal_manure_excretions import AnimalManureExcretions
 
 
 def manure_calculations(
