+++ RUFAS/routines/animal/pen.py
@@ -17,7 +17,7 @@ from RUFAS.routines.animal.manure.general_manure import (
 )
 from RUFAS.routines.animal.ration.animal_requirements import AnimalRequirements
 
-from ...data_structures.pen_manure_data import PenManureData
+from ...data_types.pen_manure_data import PenManureData
 from ...enums import AnimalCombination
 from .ration.amino_acid import EssentialAminoAcidRequirements
