+++ RUFAS/routines/manure/manure_manager.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import typing
 from typing import Any, Dict, List, Optional, Tuple
 
-from RUFAS.data_structures.pen_manure_data import PenManureData
+from RUFAS.data_types.pen_manure_data import PenManureData
 from RUFAS.output_manager import OutputManager
 from RUFAS.routines.manure.beddings.bedding_classes import BaseBedding, BeddingFactory
 from RUFAS.routines.manure.constants_and_units.manure_constants import ManureConstants
