+++ RUFAS/routines/manure/pen_manure/pen_manure.py
@@ -7,7 +7,7 @@ from RUFAS.general_constants import GeneralConstants
 from RUFAS.routines.manure.constants_and_units.manure_constants import ManureConstants
 from RUFAS.units import MeasurementUnits
 
-from ....data_structures.animal_manure_excretions import AnimalManureExcretions
+from ....data_types.animal_manure_excretions import AnimalManureExcretions
 
 
 @dataclass
