+++ tests/test_manure_module/test_animal_manure_excretions/test_calf_manure_excretion.py
@@ -2,7 +2,7 @@ import pytest
 from pytest import approx
 from pytest_mock import MockFixture
 
-from RUFAS.data_structures.animal_manure_excretions import AnimalManureExcretions
+from RUFAS.data_types.animal_manure_excretions import AnimalManureExcretions
 from RUFAS.general_constants import GeneralConstants
 from RUFAS.routines.animal.manure.calf_manure_excretion import manure_calculations
