+++ tests/test_simulation_engine.py
@@ -2,7 +2,7 @@ import pytest
 from mock.mock import MagicMock
 from pytest_mock import MockerFixture
 
-from RUFAS.data_structures.pen_manure_data import PenManureData
+from RUFAS.data_types.pen_manure_data import PenManureData
 from RUFAS.output_manager import OutputManager
 from RUFAS.routines import Feed
 from RUFAS.routines.EEE.EEE_manager import EEEManager
